Inline 'sublime_linter_lint'
