+++ sublime_linter.py
@@ -173,6 +173,38 @@ def has_syntax_changed(view):
         buffer_syntaxes[bid] = current_syntax
 
 
+class SublimeLinterLintCommand(sublime_plugin.TextCommand):
+    """A command that lints the current view if it has a linter."""
+
+    def is_enabled(self):
+        """
+        Return True if the current view can be linted.
+
+        If the view has *only* file-only linters, it can be linted
+        only if the view is not dirty.
+
+        Otherwise it can be linted.
+        """
+        has_non_file_only_linter = False
+
+        bid = self.view.buffer_id()
+        linters = persist.view_linters.get(bid, [])
+
+        for lint in linters:
+            if lint.tempfile_suffix != '-':
+                has_non_file_only_linter = True
+                break
+
+        if not has_non_file_only_linter:
+            return not self.view.is_dirty()
+
+        return True
+
+    def run(self, edit):
+        """Lint the current view."""
+        hit(self.view)
+
+
 class SublimeLinterConfigChanged(sublime_plugin.ApplicationCommand):
     def run(self):
         lint_all_views()
